perf(api): split the two dbtest files the database wave waits on
node:test parallelises across files and never within one, so a file's
serial total is a floor the whole wave sits on however many lanes it has.
Measured on the gate worker at twelve lanes, the wave was 71s against 505s
of test time — 42s of ideal packing — because two files were 54s and 45s
on their own.

migration.dbtest.ts is one cheap file plus five chain-layer migration
proofs that each replay the real migration history; those five were 45 of
its 54 seconds. They move to migration-chain-layer-expand.dbtest.ts and
migration-chain-layer-contract.dbtest.ts, with the fixture they share
extracted to chain-layer-migration-fixture.ts.

preflight-goal-execution.dbtest.ts was 45s, 27 of them in a single test
looping over four abort fixtures, each restaging the pre-kernel history.
The loop becomes four tests in preflight-goal-execution-abort.dbtest.ts,
sharing the staging harness through goal-execution-fixture.ts. One test per
case also makes node:test name the fixture that failed rather than the loop
that contained it.

No test body changed and none was dropped: 28 tests across the five files,
all green against a throwaway PostgreSQL with the gate's own settings.
