+++ packages/api/src/migration.dbtest.ts
@@ -8,148 +8,19 @@ import {
 } from "node:fs";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
-import { fileURLToPath } from "node:url";
 import { after, before, beforeEach, test } from "node:test";
 
 import { backfillTaskSource, backfilledFireId, PrismaClient } from "@anneal/db";
 
+import {
+  dbDirectory,
+  migrationHarnessEnabled,
+  migrationQuery,
+  retiredFollowUpColumn,
+  retiredFollowUpIndex,
+} from "./chain-layer-migration-fixture.js";
 import { fireCronTask } from "./scheduler.js";
 import { resetTestDb, setupTestDb, testDatabaseSchema, testDatabaseUrl } from "./testdb.js";
-
-const chainLayerExpandMigration = "20260823100000_chain_layer_expand";
-const chainLayerContractMigration = "20260824100000_chain_layer_contract";
-const retiredFollowUpColumn = ["follow", "UpTaskId"].join("");
-const retiredFollowUpIndex = `Task_${retiredFollowUpColumn}_key`;
-const dbDirectory = fileURLToPath(new URL("../../db", import.meta.url));
-
-interface ChainLayerMigrationFixture {
-  schema: string;
-  url: string;
-  quotedSchema: string;
-  execute(sql: string): void;
-  applyExpandMigration(): void;
-  applyContractMigration(): void;
-  cleanup(): void;
-}
-
-/**
- * Stage the real migration history immediately before the expand migration.
- * The fixture copies the committed Prisma tree into a temporary directory,
- * removes this migration and every later one, then adds this migration back
- * only when the test is ready to exercise it. A unique schema derived from
- * the explicitly opted-in test URL keeps the fixture away from live data.
- */
-const stageBeforeChainLayerExpand = (): ChainLayerMigrationFixture => {
-  const base = new URL(testDatabaseUrl);
-  const sourceSchema = base.searchParams.get("schema");
-  if (!sourceSchema || sourceSchema === "public") throw new Error("chain-layer migration fixture refuses public schema");
-  const schema = `agentos_chain_layer_${process.pid}_${Date.now().toString(36)}`;
-  base.searchParams.set("schema", schema);
-  const url = base.toString();
-  const quotedSchema = `"${schema.replaceAll('"', '""')}"`;
-  const execute = (sql: string): void => {
-    execFileSync("npx", ["prisma", "db", "execute", "--url", url, "--stdin"], {
-      cwd: dbDirectory,
-      input: sql,
-      stdio: ["pipe", "pipe", "pipe"],
-    });
-  };
-
-  execute(`DROP SCHEMA IF EXISTS ${quotedSchema} CASCADE; CREATE SCHEMA ${quotedSchema};`);
-  const staging = mkdtempSync(join(tmpdir(), "chain-layer-expand-fixture."));
-  cpSync(join(dbDirectory, "prisma"), join(staging, "prisma"), { recursive: true });
-  for (const entry of readdirSync(join(staging, "prisma", "migrations"), { withFileTypes: true })) {
-    if (entry.isDirectory() && entry.name >= chainLayerExpandMigration) {
-      rmSync(join(staging, "prisma", "migrations", entry.name), { recursive: true, force: true });
-    }
-  }
-
-  const deploy = (): void => {
-    execFileSync("npx", ["prisma", "migrate", "deploy", "--schema", join(staging, "prisma", "schema.prisma")], {
-      cwd: dbDirectory,
-      env: { ...process.env, DATABASE_URL: url },
-      stdio: ["ignore", "pipe", "pipe"],
-    });
-  };
-  deploy();
-
-  return {
-    schema,
-    url,
-    quotedSchema,
-    execute,
-    applyExpandMigration: () => {
-      cpSync(
-        join(dbDirectory, "prisma", "migrations", chainLayerExpandMigration),
-        join(staging, "prisma", "migrations", chainLayerExpandMigration),
-        { recursive: true },
-      );
-      deploy();
-    },
-    applyContractMigration: () => {
-      cpSync(
-        join(dbDirectory, "prisma", "migrations", chainLayerContractMigration),
-        join(staging, "prisma", "migrations", chainLayerContractMigration),
-        { recursive: true },
-      );
-      deploy();
-    },
-    cleanup: () => {
-      rmSync(staging, { recursive: true, force: true });
-      try {
-        execute(`DROP SCHEMA IF EXISTS ${quotedSchema} CASCADE;`);
-      } catch {
-        // A failed fixture must not leave its private schema behind if the
-        // server has already terminated the connection.
-      }
-    },
-  };
-};
-
-const migrationQuery = async <T>(fixture: { url: string }, sql: string): Promise<T[]> => {
-  const client = new PrismaClient({ datasources: { db: { url: fixture.url } } });
-  try {
-    return await client.$queryRawUnsafe<T[]>(sql);
-  } finally {
-    await client.$disconnect();
-  }
-};
-
-const migrationSnapshot = async (fixture: ChainLayerMigrationFixture): Promise<{ steps: string[]; tasks: string[] }> => ({
-  steps: (await migrationQuery<{ row: string }>(fixture,
-    'SELECT row_to_json(step)::text AS row FROM "TaskTemplateStep" AS step ORDER BY step."id"'))
-    .map(({ row }) => row),
-  tasks: (await migrationQuery<{ row: string }>(fixture,
-    'SELECT row_to_json(task)::text AS row FROM "Task" AS task ORDER BY task."id"'))
-    .map(({ row }) => row),
-});
-
-const migrationColumns = async (fixture: ChainLayerMigrationFixture): Promise<string[]> => (
-  await migrationQuery<{ column_name: string }>(fixture, `
-    SELECT column_name
-    FROM information_schema.columns
-    WHERE table_schema = '${fixture.schema.replaceAll("'", "''")}'
-      AND (table_name, column_name) IN (
-        ('TaskTemplateStep', 'layer'),
-        ('Task', 'chainLayer')
-      )
-    ORDER BY table_name, column_name
-  `)
-).map(({ column_name }) => column_name);
-
-const migrationFailureOutput = (error: unknown): string => {
-  const candidate = error as { stdout?: string | Buffer; stderr?: string | Buffer; message?: string };
-  return [candidate.stdout, candidate.stderr, candidate.message]
-    .filter((value): value is string | Buffer => value !== undefined)
-    .map((value) => value.toString())
-    .join("\n");
-};
-
-const migrationHarnessEnabled = (
-  process.env.AGENTOS_ALLOW_SCRATCH_DATABASES === "1"
-  && Boolean(process.env.TEST_DATABASE_URL)
-  && Boolean(process.env.TEST_DATABASE_MAINTENANCE_URL)
-);
 
 const taskDispatchBindingMigration = "20260825120000_task_dispatch_binding";
 
@@ -607,197 +478,6 @@ test("final contract accepts standalone tasks and requires complete chain identi
     `),
     /violates check constraint|Task_chain_identity_all_or_none_check/u,
   );
-});
-
-test("contract migration preserves a consistent legacy chain and removes follow-ups", {
-  skip: !migrationHarnessEnabled,
-}, async () => {
-  const fixture = stageBeforeChainLayerExpand();
-  try {
-    fixture.execute(`
-      INSERT INTO "Project" ("id", "name", "slug", "updatedAt")
-      VALUES ('contract-project', 'contract-project', 'contract-project', NOW());
-      INSERT INTO "TaskTemplate" ("id", "projectId", "name", "description", "variables", "updatedAt")
-      VALUES ('contract-template', 'contract-project', 'legacy-template', 'legacy', ARRAY[]::text[], NOW());
-      INSERT INTO "TaskTemplateStep" ("id", "taskTemplateId", "stepIndex", "name", "assigneeType", "prompt")
-      VALUES ('contract-step-1', 'contract-template', 1, 'step 1', 'agent', 'prompt 1'),
-             ('contract-step-2', 'contract-template', 2, 'step 2', 'agent', 'prompt 2');
-      INSERT INTO "Task" ("id", "projectId", "name", "description", "chainId", "chainIndex", "updatedAt")
-      VALUES ('contract-task-1', 'contract-project', 'task 1', 'task 1', 'contract-chain', 1, NOW()),
-             ('contract-task-2', 'contract-project', 'task 2', 'task 2', 'contract-chain', 2, NOW());
-      UPDATE "Task" SET "${retiredFollowUpColumn}" = 'contract-task-2' WHERE "id" = 'contract-task-1';
-    `);
-    fixture.applyExpandMigration();
-    fixture.applyContractMigration();
-
-    assert.deepEqual(
-      await migrationQuery<{ stepIndex: number; layer: number }>(fixture,
-        'SELECT "stepIndex", "layer" FROM "TaskTemplateStep" ORDER BY "stepIndex"'),
-      [{ stepIndex: 1, layer: 1 }, { stepIndex: 2, layer: 2 }],
-    );
-    assert.deepEqual(
-      await migrationQuery<{ chainIndex: number; chainLayer: number }>(fixture,
-        'SELECT "chainIndex", "chainLayer" FROM "Task" WHERE "chainId" = \'contract-chain\' ORDER BY "chainIndex"'),
-      [{ chainIndex: 1, chainLayer: 1 }, { chainIndex: 2, chainLayer: 2 }],
-    );
-    assert.deepEqual(await migrationColumns(fixture), ["chainLayer", "layer"]);
-    assert.deepEqual(
-      await migrationQuery<{ indexname: string }>(fixture,
-        `SELECT indexname FROM pg_indexes WHERE schemaname = current_schema() AND indexname = '${retiredFollowUpIndex}'`),
-      [],
-    );
-  } finally {
-    fixture.cleanup();
-  }
-});
-
-test("contract migration refuses an inconsistent follow-up before tightening or dropping", {
-  skip: !migrationHarnessEnabled,
-}, async () => {
-  const fixture = stageBeforeChainLayerExpand();
-  try {
-    fixture.execute(`
-      INSERT INTO "Project" ("id", "name", "slug", "updatedAt")
-      VALUES ('contract-fence-project', 'contract-fence-project', 'contract-fence-project', NOW());
-      INSERT INTO "TaskTemplate" ("id", "projectId", "name", "description", "variables", "updatedAt")
-      VALUES ('contract-fence-template', 'contract-fence-project', 'legacy-template', 'legacy', ARRAY[]::text[], NOW());
-      INSERT INTO "TaskTemplateStep" ("id", "taskTemplateId", "stepIndex", "name", "assigneeType", "prompt")
-      VALUES ('contract-fence-step', 'contract-fence-template', 1, 'step', 'agent', 'prompt');
-      INSERT INTO "Task" ("id", "projectId", "name", "description", "chainId", "chainIndex", "updatedAt")
-      VALUES ('contract-fence-source', 'contract-fence-project', 'source', 'source', 'contract-fence-chain', 1, NOW()),
-             ('contract-fence-target', 'contract-fence-project', 'target', 'target', 'contract-fence-chain', 2, NOW());
-    `);
-    fixture.applyExpandMigration();
-    // The expand fence has already run. Introduce a legacy inconsistency after
-    // it so this test exercises the contract migration's second fence.
-    fixture.execute(`UPDATE "Task" SET "${retiredFollowUpColumn}" = 'contract-fence-source' WHERE "id" = 'contract-fence-target';`);
-    const before = await migrationSnapshot(fixture);
-    let error: unknown;
-    try {
-      fixture.applyContractMigration();
-    } catch (caught) {
-      error = caught;
-    }
-    assert.ok(error, "the contract follow-up consistency fence must fail");
-    assert.match(migrationFailureOutput(error), /chain-layer-contract: inconsistent-follow-up-relationship/u);
-    assert.deepEqual(await migrationSnapshot(fixture), before);
-    assert.deepEqual(await migrationColumns(fixture), ["chainLayer", "layer"]);
-    assert.deepEqual(
-      await migrationQuery<{ is_nullable: string }>(fixture,
-        `SELECT is_nullable FROM information_schema.columns WHERE table_schema = '${fixture.schema.replaceAll("'", "''")}' AND table_name = 'TaskTemplateStep' AND column_name = 'layer'`),
-      [{ is_nullable: "YES" }],
-    );
-    assert.deepEqual(
-      await migrationQuery<{ column_name: string }>(fixture,
-        `SELECT column_name FROM information_schema.columns WHERE table_name = 'Task' AND column_name = '${retiredFollowUpColumn}'`),
-      [{ column_name: retiredFollowUpColumn }],
-    );
-  } finally {
-    fixture.cleanup();
-  }
-});
-
-test("chain-layer expand migration dense-ranks legacy template steps and chain nodes", {
-  skip: !migrationHarnessEnabled,
-}, async () => {
-  const fixture = stageBeforeChainLayerExpand();
-  try {
-    fixture.execute(`
-      INSERT INTO "Project" ("id", "name", "slug", "updatedAt")
-      VALUES ('chain-layer-project', 'chain-layer-project', 'chain-layer-project', NOW());
-      INSERT INTO "TaskTemplate" ("id", "projectId", "name", "description", "variables", "updatedAt")
-      VALUES ('chain-layer-template', 'chain-layer-project', 'legacy-template', 'legacy', ARRAY[]::text[], NOW());
-      INSERT INTO "TaskTemplateStep" ("id", "taskTemplateId", "stepIndex", "name", "assigneeType", "prompt")
-      VALUES
-        ('chain-layer-step-09', 'chain-layer-template', 9, 'step 9', 'agent', 'prompt'),
-        ('chain-layer-step-40', 'chain-layer-template', 40, 'step 40', 'agent', 'prompt'),
-        ('chain-layer-step-90', 'chain-layer-template', 90, 'step 90', 'agent', 'prompt');
-      INSERT INTO "Task" ("id", "projectId", "name", "description", "chainId", "chainIndex", "updatedAt")
-      VALUES
-        ('chain-layer-task-08', 'chain-layer-project', 'task 8', 'task', 'legacy-chain', 8, NOW()),
-        ('chain-layer-task-40', 'chain-layer-project', 'task 40', 'task', 'legacy-chain', 40, NOW()),
-        ('chain-layer-task-90', 'chain-layer-project', 'task 90', 'task', 'legacy-chain', 90, NOW());
-    `);
-
-    fixture.applyExpandMigration();
-
-    assert.deepEqual(
-      await migrationQuery<{ stepIndex: number; layer: number | null }>(fixture,
-        'SELECT "stepIndex", "layer" FROM "TaskTemplateStep" ORDER BY "stepIndex"'),
-      [
-        { stepIndex: 9, layer: 1 },
-        { stepIndex: 40, layer: 2 },
-        { stepIndex: 90, layer: 3 },
-      ],
-    );
-    assert.deepEqual(
-      await migrationQuery<{ chainIndex: number; chainLayer: number | null }>(fixture,
-        'SELECT "chainIndex", "chainLayer" FROM "Task" WHERE "chainId" = \'legacy-chain\' ORDER BY "chainIndex"'),
-      [
-        { chainIndex: 8, chainLayer: 1 },
-        { chainIndex: 40, chainLayer: 2 },
-        { chainIndex: 90, chainLayer: 3 },
-      ],
-    );
-  } finally {
-    fixture.cleanup();
-  }
-});
-
-test("partial chain identity aborts expand before changing rows", {
-  skip: !migrationHarnessEnabled,
-}, async () => {
-  const fixture = stageBeforeChainLayerExpand();
-  try {
-    fixture.execute(`
-      INSERT INTO "Project" ("id", "name", "slug", "updatedAt")
-      VALUES ('partial-project', 'partial-project', 'partial-project', NOW());
-      INSERT INTO "Task" ("id", "projectId", "name", "description", "chainId", "updatedAt")
-      VALUES ('partial-task', 'partial-project', 'partial', 'partial', 'partial-chain', NOW());
-    `);
-    const before = await migrationSnapshot(fixture);
-    let error: unknown;
-    try {
-      fixture.applyExpandMigration();
-    } catch (caught) {
-      error = caught;
-    }
-    assert.ok(error, "the partial-chain preflight must fail");
-    assert.match(migrationFailureOutput(error), /chain-layer-expand: partial-chain-identity/u);
-    assert.deepEqual(await migrationSnapshot(fixture), before);
-    assert.deepEqual(await migrationColumns(fixture), []);
-  } finally {
-    fixture.cleanup();
-  }
-});
-
-test("inconsistent follow-up relationship aborts expand before changing rows", {
-  skip: !migrationHarnessEnabled,
-}, async () => {
-  const fixture = stageBeforeChainLayerExpand();
-  try {
-    fixture.execute(`
-      INSERT INTO "Project" ("id", "name", "slug", "updatedAt")
-      VALUES ('follow-up-project', 'follow-up-project', 'follow-up-project', NOW());
-      INSERT INTO "Task" ("id", "projectId", "name", "description", "chainId", "chainIndex", "updatedAt")
-      VALUES ('follow-up-target', 'follow-up-project', 'target', 'target', 'chain-b', 2, NOW());
-      INSERT INTO "Task" ("id", "projectId", "name", "description", "chainId", "chainIndex", "${retiredFollowUpColumn}", "updatedAt")
-      VALUES ('follow-up-source', 'follow-up-project', 'source', 'source', 'chain-a', 1, 'follow-up-target', NOW());
-    `);
-    const before = await migrationSnapshot(fixture);
-    let error: unknown;
-    try {
-      fixture.applyExpandMigration();
-    } catch (caught) {
-      error = caught;
-    }
-    assert.ok(error, "the follow-up consistency preflight must fail");
-    assert.match(migrationFailureOutput(error), /chain-layer-expand: inconsistent-follow-up-relationship/u);
-    assert.deepEqual(await migrationSnapshot(fixture), before);
-    assert.deepEqual(await migrationColumns(fixture), []);
-  } finally {
-    fixture.cleanup();
-  }
 });
 
 test("dispatch binding migration is additive and preserves existing task rows", {
