+++ packages/api/src/preflight-goal-execution.dbtest.ts
@@ -1,9 +1,7 @@
 import assert from "node:assert/strict";
-import { execFileSync } from "node:child_process";
-import { fileURLToPath } from "node:url";
 import { after, test } from "node:test";
 
-import { preKernelRun, preKernelSeed, stageAtPreviousMigration } from "./goal-execution-fixture.js";
+import { preflightHarness, preKernelRun } from "./goal-execution-fixture.js";
 
 /**
  * The Goal 5a0 migration preflight, exercised as the operator runs it.
@@ -12,51 +10,22 @@ import { preKernelRun, preKernelSeed, stageAtPreviousMigration } from "./goal-ex
  * must abort *before* the migration, and the report must carry IDs and counts
  * and nothing else. It runs against a database staged at the migration before
  * the kernel, because that is the only state in which the preflight is ever run.
+ *
+ * The four abort fixtures live in preflight-goal-execution-abort.dbtest.ts.
+ * Each of them restages the whole pre-kernel history, and the tests in one file
+ * run one after another, so together they were the longest single file in the
+ * database wave.
  */
 
-const dbDirectory = fileURLToPath(new URL("../../db", import.meta.url));
-
-let fixture: ReturnType<typeof stageAtPreviousMigration>;
+const harness = preflightHarness("preflight");
 after(() => {
-  fixture?.cleanup();
+  harness.cleanup();
 });
 
-const stage = (rows: string[]): void => {
-  fixture?.cleanup();
-  fixture = stageAtPreviousMigration("preflight");
-  fixture.execute(preKernelSeed + rows.join("\n"));
-};
-
-interface PreflightResult { code: number; stdout: string; stderr: string }
-
-const runPreflight = (environment: Record<string, string | undefined> = {}): PreflightResult => {
-  const env = {
-    ...process.env,
-    DATABASE_URL: fixture.url,
-    PATH: `${dbDirectory}/node_modules/.bin:${process.env.PATH ?? ""}`,
-    ...environment,
-  } as NodeJS.ProcessEnv;
-  for (const [key, value] of Object.entries(environment)) if (value === undefined) delete env[key];
-  try {
-    // Run exactly as `npm run db:preflight-goal-execution` runs it: from the db
-    // workspace, not from the repository root, so the script resolves its own
-    // dependencies exactly as the wired script does.
-    const stdout = execFileSync("npx", ["tsx", "prisma/preflight-goal-execution.ts"], {
-      cwd: dbDirectory,
-      env,
-      encoding: "utf8",
-    });
-    return { code: 0, stdout, stderr: "" };
-  } catch (error) {
-    const failure = error as { status?: number; stdout?: string; stderr?: string };
-    return { code: failure.status ?? -1, stdout: String(failure.stdout ?? ""), stderr: String(failure.stderr ?? "") };
-  }
-};
-
 test("a clean history passes and the report carries only IDs and counts", () => {
-  stage([preKernelRun("r-1", "t-old", "g-up", 1)]);
+  harness.stage([preKernelRun("r-1", "t-old", "g-up", 1)]);
 
-  const result = runPreflight();
+  const result = harness.runPreflight();
   assert.equal(result.code, 0, `${result.stdout}\n${result.stderr}`);
   assert.match(result.stdout, /preflight PASS/u);
   assert.match(result.stdout, /preflight count tasksToBackfill=1/u);
@@ -64,43 +33,8 @@ test("a clean history passes and the report carries only IDs and counts", () => 
   assert.doesNotMatch(result.stdout + result.stderr, /secret-looking/u, "no prompt or spec text reaches the report");
 });
 
-test("every ambiguous, corrupt, or active fixture aborts before the migration", () => {
-  const cases: Array<{ label: string; condition: string; rows: string[]; extra?: string }> = [
-    {
-      label: "a Task with both null and non-null Run.goalId",
-      condition: "mixed-lineage",
-      rows: [preKernelRun("r-a", "t-old", "g-up", 1), preKernelRun("r-b", "t-old", null, 2)],
-    },
-    {
-      label: "a Task whose Runs name different Goals",
-      condition: "ambiguous-goal",
-      extra: `INSERT INTO "Goal" ("id", "projectId", "title", "spec", "updatedAt")
-              VALUES ('g-second', 'p-up', 'Second', 'spec', NOW());`,
-      rows: [preKernelRun("r-a", "t-old", "g-up", 1), preKernelRun("r-b", "t-old", "g-second", 2)],
-    },
-    {
-      label: "a Goal-linked Run with no Task",
-      condition: "orphan-run",
-      rows: [preKernelRun("r-a", null, "g-up", 1)],
-    },
-    {
-      label: "an active Goal-linked Run",
-      condition: "active-run",
-      rows: [preKernelRun("r-a", "t-old", "g-up", 1, "running")],
-    },
-  ];
-
-  for (const testCase of cases) {
-    stage(testCase.extra ? [testCase.extra, ...testCase.rows] : testCase.rows);
-    const result = runPreflight();
-    assert.equal(result.code, 1, `${testCase.label} must abort: ${result.stdout}${result.stderr}`);
-    assert.match(result.stderr, new RegExp(`STOP preflight ${testCase.condition}`, "u"), testCase.label);
-    assert.doesNotMatch(result.stdout, /preflight PASS/u, testCase.label);
-  }
-});
-
 test("a Goal/Run project disagreement cannot even be created on the pre-kernel schema", () => {
-  stage([preKernelRun("r-1", "t-old", "g-up", 1)]);
+  const fixture = harness.stage([preKernelRun("r-1", "t-old", "g-up", 1)]);
   // The preflight checks this condition anyway, as defence in depth. The pre-kernel
   // composite foreign key Run(goalId, projectId) -> Goal already makes the corrupt
   // state unreachable, and that is worth pinning: if a future migration weakens the
@@ -116,11 +50,11 @@ test("a Goal/Run project disagreement cannot even be created on the pre-kernel s
 });
 
 test("an unnamed schema stops the preflight", () => {
-  stage([preKernelRun("r-1", "t-old", "g-up", 1)]);
+  const fixture = harness.stage([preKernelRun("r-1", "t-old", "g-up", 1)]);
 
   const url = new URL(fixture.url);
   url.searchParams.delete("schema");
-  const noSchema = runPreflight({ DATABASE_URL: url.toString() });
+  const noSchema = harness.runPreflight({ DATABASE_URL: url.toString() });
   assert.equal(noSchema.code, 1);
   assert.match(noSchema.stderr, /must name the target schema explicitly/u);
 });
